+++ packages/common/src/mvc/interfaces/index.ts
@@ -3,3 +3,4 @@ export * from "./PathParamsType";
 export * from "./ResponseErrorObject";
 export * from "./IParamOptions";
 export * from "./HandlerType";
+export * from "./PlatformRouteOptions";
