+++ packages/common/src/platform/utils/createHandlerMetadata.ts
@@ -1,8 +1,7 @@
 import {InjectorService} from "@tsed/di";
-import {HandlerType} from "../../mvc/interfaces/HandlerType";
+import {HandlerType, PlatformRouteWithoutHandlers} from "../../mvc/interfaces";
 import {EndpointMetadata} from "../../mvc/models/EndpointMetadata";
 import {HandlerMetadata, HandlerMetadataOptions} from "../../mvc/models/HandlerMetadata";
-import type {PlatformRouteWithoutHandlers} from "../interfaces/PlatformRouterMethods";
 
 function isMetadata(input: any) {
   return input instanceof HandlerMetadata;
